+++ java-lazy/src/main/java/solution/andrei/city/equipment/model/Kettle.java
@@ -1,8 +1,8 @@
 package solution.andrei.city.equipment.model;
 
 public class Kettle extends BaseElectricEquipment {
-    private final int capacity;
-    public Kettle(int power, int capacity) {
+    private final double capacity;
+    public Kettle(int power, double capacity) {
         super(EquipmentType.REGULAR, power);
         this.capacity = capacity;
     }
